Refactor: Apply Coroutine, Divide View, ViewModel, Repository SignUp 관련 Repository 생성 UserViewModel 생성 View에서 DB 접근 -> ViewModel -> Repository로 이동 (MVVM Pattern 적용) Coroutine 적용
