+++ app/build.gradle
@@ -56,6 +56,7 @@ dependencies {
     implementation 'androidx.annotation:annotation:1.3.0'
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.4.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.0'
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.4.0"
     implementation 'com.google.firebase:firebase-firestore-ktx:24.0.0'
     implementation 'com.google.firebase:firebase-storage-ktx:20.0.0'
     implementation 'com.android.volley:volley:1.2.0'
@@ -108,8 +109,6 @@ dependencies {
     //firebase 푸시알람 라이브러리
     implementation 'com.google.firebase:firebase-messaging:17.3.4'
 
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
     implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
 
     implementation 'androidx.activity:activity-ktx:1.2.0'
@@ -129,7 +128,6 @@ dependencies {
     // Testing Fragments in Isolation
     debugImplementation("androidx.fragment:fragment-testing:$fragment_version")
     implementation "androidx.core:core-ktx:+"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     //pull and refresh
